fix : Refactor StackViewport to use adjustInitialViewUp utility for view up adjustments and update Playwright config command for cross-env compatibility
